Compute energy of all spatially-filtered test data
This is used to compute the boxplots of the energy of the preictal and
interictal data using each CSP filter, for all the spectral bands. Then,
the area under the (ROC or precision-recall) curve is computed to find
the spectral band with highest classification performance. In "Searching
for waveforms on spatially-filtered epileptic ECoG", we used the ROC
curve. In "Shift-invariant waveform learning on epileptic ECoG", we
realized that the problem is highly class imbalanced (we fixed the ratio
of preictal:interictal data to 1:3), so we used the precision-recall
curve, as it bahaves better on imbalanced problems.

In this version we also compute and save the full matrix of CSP filters.
We might want to use more than one filter per class in future works.